chg: Learning Kubernetes
Go back to using 2 CPUs for the control plane nodes in the Vagrantfile.
The reason for this is that Antrea by default take up 400m CPU,
combined with the 650m that the control-plane uses puts us over the 1
CPU limit.

Workaround would be to modify the Antrea deployment to use less CPU,
say 100m for the Deployment and DaemonSet.